Cleaned up the pipeline nodes, estimators, transformer interfaces
